requirements.txt
flask
flask-sqlalchemy
celery
redis
pillow
requests
python-dotenv
psycopg2-binary
flask-cors
pandas